Fixed queries - Code beautify
