+++ index.js
@@ -1,52 +1,10 @@
 //dependencies initialization
-const express = require('express');
-const bodyParser = require("body-parser");
-const sql = require("mssql");
+const sql = require("./sql");
+const app = require("./webserver");
 
-function passValue(value) {
-    return 'JSON.parse(Base64.decode("' + new Buffer(JSON.stringify(value)).toString('base64') + '"))'
-  }
-
-// sqlconfig for your database - Come separare questa cosa?
-var sqlconfig = {
-    server: 'htc-demo.database.windows.net', 
-    database: 'htc-demo',
-    authentication: {
-        type: 'azure-active-directory-default',//DA RIVEDERE
-        options: {
-            clientId: "ef7021e1-407c-43b6-a72d-d7d5181fa504"
-        }
-    },
-    trustServerCertificate: false //self-signed cert error
-};
-
-// connect to your database
-sql.connect(sqlconfig, function (err) {
-    if (err){
-        throw err;
-    } 
-    console.log(`Connected to DB: ${sqlconfig.server}`);
-});
-
-//webserver initialization
-const app = express();              
-const port = 80;
-
-//webserver listener and registrations
-app.listen(port, () => { console.log(`Now listening on port ${port}`); }); //webserver listener
-app.set('view engine', 'ejs');
-
-app.use(express.static(__dirname));
-app.use(bodyParser.urlencoded({ extended: true }))
-
-//get & post methods
+//WebServer GET & POST Methods
 app.get('/', (req, res) => {
     res.render('index');      
-});
-
-app.post("/", (req, res) => {
-    console.log(req.body);
-    res.render('index');
 });
 
 app.post("/register", (req, res) => {
@@ -84,7 +42,7 @@ app.post("/register", (req, res) => {
 
 app.post("/requests", (req, res) =>{
     var request = new sql.Request();
-    var query = `SELECT * FROM [SalesLT].Contact`;
+    var query = "SELECT * FROM [SalesLT].Contact";
     request.query(query, (err, recordset) => {
         if (err){
             console.log("Error: " + err)
@@ -94,17 +52,17 @@ app.post("/requests", (req, res) =>{
     })
 })
 
-app.get("/requests",(req, res) =>{
+// //TEST DEV REFERENCE `'"
 
-    console.log(html);
-    res.send(html);
-})
+// app.get("/requests",(req, res) =>{
 
-//TEST DEV REFERENCE `
+//     console.log(html);
+//     res.send(html);
+// })
 
-app.get('/register', (req, res) => {
-    res.render('register', {name: "notset", email: "notset", subject: "notset", message: "notset"});     
-});
-app.get('/register', (req, res) => {
-    res.render('register', {name: "notset", email: "notset", subject: "notset", message: "notset"});     
-});
+// app.get('/register', (req, res) => {
+//     res.render('register', {name: "notset", email: "notset", subject: "notset", message: "notset"});     
+// });
+// app.get('/register', (req, res) => {
+//     res.render('register', {name: "notset", email: "notset", subject: "notset", message: "notset"});     
+// });
