+++ package.json
@@ -24,9 +24,11 @@
   "homepage": "https://github.com/crypto-infinity/inft_gui#readme",
   "dependencies": {
     "body-parser": "^1.20.2",
+    "dotenv": "^16.3.1",
     "ejs": "^3.1.9",
     "express": "^4.18.2",
     "knex": "^2.5.1",
+    "linkedin-private-api": "^1.1.2",
     "mssql": "^10.0.1",
     "pug": "^3.0.2"
   }
